fix: separate tests into correct file
