+++ internal/auth/auth_test.go
@@ -1,0 +1,62 @@
+package auth
+
+import (
+	"net/http"
+	"testing"
+)
+
+func TestGetAPIKey(t *testing.T) {
+	tests := []struct {
+		name          string
+		headers       http.Header
+		expectedKey   string
+		expectedError error
+	}{
+		{
+			name: "Valid API Key",
+			headers: http.Header{
+				"Authorization": []string{"ApiKey my-secret-api-key"},
+			},
+			expectedKey:   "my-secret-api-key",
+			expectedError: nil,
+		},
+		{
+			name:          "No Authorization Header",
+			headers:       http.Header{},
+			expectedKey:   "",
+			expectedError: ErrNoAuthHeaderIncluded,
+		},
+		{
+			name: "Malformed Authorization Header",
+			headers: http.Header{
+				"Authorization": []string{"Bearer my-secret-api-key"},
+			},
+			expectedKey:   "",
+			expectedError: nil,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			key, err := GetAPIKey(tc.headers)
+			if tc.expectedError != nil {
+				if err != tc.expectedError {
+					t.Errorf("expected error %v, got %v", tc.expectedError, err)
+				}
+				return
+			}
+			if tc.name == "Malformed Authorization Header" {
+				if err == nil {
+					t.Error("expected an error for malformed header, but got none")
+				}
+				return
+			}
+			if err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+			if key != tc.expectedKey {
+				t.Errorf("expected key %v, got %v", tc.expectedKey, key)
+			}
+		})
+	}
+}
